requirements.txt
openai
ipykernel
python-dotenv
streamlit